create segmentation keys enum
