+++ flash/core/classification.py
@@ -14,9 +14,6 @@
 from dataclasses import dataclass
 from typing import Any, Callable, Dict, List, Mapping, Optional, Sequence, Tuple, Union
 
-# for visualisation
-import kornia as K
-import matplotlib.pyplot as plt
 import torch
 import torch.nn.functional as F
 import torchmetrics
@@ -167,55 +164,3 @@ class Labels(Classes):
                 "No ClassificationState was found, this serializer will act as a Classes serializer.", UserWarning
             )
             return classes
-
-
-class SegmentationLabels(Serializer):
-
-    def __init__(self, labels_map: Optional[Dict[int, Tuple[int, int, int]]] = None, visualise: bool = False):
-        """A :class:`.Serializer` which converts the model outputs to the label of the argmax classification
-        per pixel in the image for semantic segmentation tasks.
-
-        Args:
-            labels_map: A dictionary that map the labels ids to pixel intensities.
-            visualise: Wether to visualise the image labels.
-        """
-        super().__init__()
-        self.labels_map = labels_map
-        self.visualise = visualise
-
-    @staticmethod
-    def labels_to_image(img_labels: torch.Tensor, labels_map: Dict[int, Tuple[int, int, int]]) -> torch.Tensor:
-        """Function that given an image with labels ids and their pixels intrensity mapping,
-           creates a RGB representation for visualisation purposes.
-        """
-        assert len(img_labels.shape) == 2, img_labels.shape
-        H, W = img_labels.shape
-        out = torch.empty(3, H, W, dtype=torch.uint8)
-        for label_id, label_val in labels_map.items():
-            mask = (img_labels == label_id)
-            for i in range(3):
-                out[i].masked_fill_(mask, label_val[i])
-        return out
-
-    @staticmethod
-    def create_random_labels_map(num_classes: int) -> Dict[int, Tuple[int, int, int]]:
-        labels_map: Dict[int, Tuple[int, int, int]] = {}
-        for i in range(num_classes):
-            labels_map[i] = torch.randint(0, 255, (3, ))
-        return labels_map
-
-    def serialize(self, sample: torch.Tensor) -> torch.Tensor:
-        assert len(sample.shape) == 3, sample.shape
-        labels = torch.argmax(sample, dim=-3)  # HxW
-        if self.visualise:
-            if self.labels_map is None:
-                # create random colors map
-                num_classes = sample.shape[-3]
-                labels_map = self.create_random_labels_map(num_classes)
-            else:
-                labels_map = self.labels_map
-            labels_vis = self.labels_to_image(labels, labels_map)
-            labels_vis = K.utils.tensor_to_image(labels_vis)
-            plt.imshow(labels_vis)
-            plt.show()
-        return labels
